backend/requirements.txt
langgraph
langchain
langchain-openai
langchain-community
fastapi
uvicorn
python-dotenv
httpx
google-auth
google-api-python-client
apscheduler
pydantic-settings
